+++ src/main/java/com/rgosiewski/frameiq/server/common/populator/ByIdDataPopulator.java
@@ -2,5 +2,5 @@ package com.rgosiewski.frameiq.server.common.populator;
 
 public abstract class ByIdDataPopulator<SOURCE, TARGET> {
 
-    public abstract TARGET populate(int sourceId, SOURCE source);
+    public abstract TARGET populate(Long sourceId, SOURCE source);
 }
